The session file has to be created at runtime, rather than baked into the container
The R session script seems to be run in a context that is free of
environment variables, so those imported from the host don't get set.
Moreover, R_LIBS_USER seems to be overridden by RStudio Server, so it
needs to be set explicitly at runtime and bind mounted somewhere useful.